+++ graphix/opengraph.py
@@ -17,9 +17,13 @@ from graphix.measurements import BlochMeasurement, Measurement
 if TYPE_CHECKING:
     from collections.abc import Callable, Collection, Iterable, Mapping, Sequence
 
+    # Unpack introduced in Python 3.12
+    from typing_extensions import Unpack
+
     from graphix.flow.core import CausalFlow
     from graphix.parameter import ExpressionOrSupportsFloat, Parameter
     from graphix.pattern import Pattern
+    from graphix.visualization import DrawKwargs
 
 # TODO: Maybe move these definitions to graphix.fundamentals and graphix.measurements ? Now they are redefined in graphix.flow._find_gpflow, not very elegant.
 _AM_co = TypeVar("_AM_co", bound=AbstractMeasurement, covariant=True)
@@ -128,6 +132,20 @@ class OpenGraph(Generic[_AM_co]):
         if flow is not None:
             return flow.to_corrections().to_pattern()
         raise OpenGraphError("The open graph does not have flow. It does not support a deterministic pattern.")
+
+    def draw(self, **options: Unpack[DrawKwargs]) -> None:
+        """Visualize the opengraph.
+
+        Parameters
+        ----------
+        options : Unpack[DrawKwargs]
+            Options controlling graph visualization. See :class:`VisualizationOptions`.
+        """
+        from graphix.visualization import GraphVisualizer  # noqa: PLC0415  Avoid circular imports
+
+        gv = GraphVisualizer.from_opengraph(og=self, **options)
+
+        gv.visualize()
 
     def map(self: OpenGraph[_A], f: Callable[[_A], _B]) -> OpenGraph[_B]:
         """Apply a function to every measurement of the open graph and return the resulting open graph.